Replace specifying per GPU memory with Accelerator spec
Take a backwards step and don't try to calculate the number of GPUs
from the needed memory.

However also replace the irritating nodeSelector requirement to
specify the per GPU memory with an accelerator interface spec.

Currently you just tell it to use CUDA and optionally the compute
version (only for documentation ATM). It then assigns one GPU and you
can override it to assign more.